+++ CHANGELOG.md
@@ -27,12 +27,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 # Past Releases
 
+## [0.33.1] - 2021-11-22
+
+### Fixed
+
+- Fixed an issue where `BlueprintView` would not size correctly when used with Auto Layout.
+
 ## [0.33.0] - 2021-11-18
 
 ### Added
 
 - Allow measuring within an explicit `SizeConstraint` in `GeometryProxy`.
-
 - Add an additional `stackLayoutChild(priority:)` method overload, for easier autocomplete when only customizing the layout priority.
 
 ### Changed
@@ -719,7 +724,8 @@ searchField
 
 - First stable release.
 
-[main]: https://github.com/square/Blueprint/compare/0.33.0...HEAD
+[main]: https://github.com/square/Blueprint/compare/0.33.1...HEAD
+[0.33.1]: https://github.com/square/Blueprint/compare/0.33.0...0.33.1
 [0.33.0]: https://github.com/square/Blueprint/compare/0.32.0...0.33.0
 [0.32.0]: https://github.com/square/Blueprint/compare/0.31.0...0.32.0
 [0.31.0]: https://github.com/square/Blueprint/compare/0.30.0...0.31.0